refactor(dependencies): Update package dependencies and restructure imports for clarity
- Removed unused dependencies from yarn.lock.
- Added new dependencies including dpdm for dependency management.
- Updated package.json files to reflect changes in dependencies and added new scripts for type and dependency checks.
- Refactored import paths in various files to improve clarity and maintainability, moving utility functions to a dedicated utils directory.
- Consolidated GIR-related utility functions to enhance type resolution and parsing capabilities.
