+++ packages/generator-typescript/package.json
@@ -46,6 +46,7 @@
     "typescript": "^5.8.3"
   },
   "dependencies": {
+    "@gi.ts/parser": "workspace:^",
     "@ts-for-gir/generator-base": "workspace:^",
     "@ts-for-gir/lib": "workspace:^",
     "ejs": "^3.1.10",
