+++ packages/lib/src/gir-module.ts
@@ -1,5 +1,4 @@
-// TODO move this class into a web-worker? https://www.npmjs.com/package/web-worker
- 
+import { type GirType } from '@gi.ts/parser'
 import { transformGirDocTagText } from './utils/index.ts'
 import { Logger } from './logger.ts'
 import { DependencyManager } from './dependency-manager.ts'
@@ -8,7 +7,6 @@ import { LibraryVersion } from './library-version.ts'
 
 import type {
     Dependency,
-    GirType,
     GirConstantElement,
     TsDocTag,
     GirInterfaceElement,
@@ -26,8 +24,9 @@ import {
     BinaryType,
     NullableType,
     ObjectType,
-    type GirNSMember,
 } from './gir.ts'
+
+import type { GirNSMember } from './gir/namespace.ts'
 import { IntrospectedAlias } from './gir/alias.ts'
 import { IntrospectedBase, IntrospectedNamespaceMember } from './gir/base.ts'
 import { IntrospectedBaseClass, IntrospectedClass, IntrospectedRecord, IntrospectedInterface } from './gir/class.ts'
@@ -41,7 +40,7 @@ import {
     IntrospectedClassFunction,
 } from './gir/function.ts'
 import { NSRegistry } from './gir/registry.ts'
-import { isPrimitiveType } from './gir/util.ts'
+import { isPrimitiveType } from './utils/index.ts'
 import { GirVisitor } from './visitor.ts'
 
 import type { OptionsLoad } from './types/index.ts'
@@ -706,62 +705,4 @@ export class GirModule {
                 .forEach((c) => this.members.set(c.name, c))
         }
     }
-}
-
-export function promisifyNamespaceFunctions(namespace: GirModule) {
-    return namespace.members.forEach((node) => {
-        if (!(node instanceof IntrospectedFunction)) return
-
-        if (node.parameters.length < 1) return
-
-        const last_param = node.parameters[node.parameters.length - 1]
-
-        if (!last_param) return
-
-        const last_param_unwrapped = last_param.type.unwrap()
-
-        if (!(last_param_unwrapped instanceof ClosureType)) return
-
-        const internal = last_param_unwrapped.type
-
-        if (internal instanceof TypeIdentifier && internal.is('Gio', 'AsyncReadyCallback')) {
-            const async_res = [
-                ...Array.from(namespace.members.values()).filter(
-                    (m): m is IntrospectedFunction => m instanceof IntrospectedFunction,
-                ),
-            ].find((m) => m.name === `${node.name.replace(/_async$/, '')}_finish` || m.name === `${node.name}_finish`)
-
-            if (async_res) {
-                const async_parameters = node.parameters.slice(0, -1).map((p) => p.copy())
-                const sync_parameters = node.parameters.map((p) => p.copy({ isOptional: false }))
-                const output_parameters = async_res.output_parameters
-
-                let async_return = new PromiseType(async_res.return())
-
-                if (output_parameters.length > 0) {
-                    const raw_return = async_res.return()
-                    if (raw_return.equals(VoidType) || raw_return.equals(BooleanType)) {
-                        const [output_type, ...output_types] = output_parameters.map((op) => op.type)
-                        async_return = new PromiseType(new TupleType(output_type, ...output_types))
-                    } else {
-                        const [...output_types] = output_parameters.map((op) => op.type)
-                        async_return = new PromiseType(new TupleType(raw_return, ...output_types))
-                    }
-                }
-
-                namespace.members.set(node.name, [
-                    node.copy({
-                        parameters: async_parameters,
-                        return_type: async_return,
-                    }),
-                    node.copy({
-                        parameters: sync_parameters,
-                    }),
-                    node.copy({
-                        return_type: new BinaryType(async_return, node.return()),
-                    }),
-                ])
-            }
-        }
-    })
 }
