+++ packages/lib/src/gir/index.ts
@@ -6,9 +6,7 @@ export * from "./enum.ts";
 export * from "./function.ts";
 export * from "./generics.ts";
 export * from "./namespace.ts";
-export * from "./nodes.ts";
 export * from "./promisify.ts";
 export * from "./property.ts";
 export * from "./registry.ts";
 export * from "./signal.ts";
-export * from "./util.ts";
