+++ packages/lib/src/gir/namespace.ts
@@ -1,7 +1,6 @@
 import { BinaryType, VoidType, PromiseType, BooleanType, TupleType, TypeIdentifier, ClosureType } from "../gir.ts";
 
-import type { GirInfoAttrs } from "../index.ts";
-import { GirModule } from "../index.ts";
+import { GirModule } from "../gir-module.ts";
 
 import type { IntrospectedBaseClass } from "./class.ts";
 import { IntrospectedFunction } from "./function.ts";
@@ -9,6 +8,7 @@ import type { IntrospectedEnum } from "./enum.ts";
 import type { IntrospectedConstant } from "./const.ts";
 import type { IntrospectedAlias } from "./alias.ts";
 
+// TODO: Move to types/
 export type GirNSMember =
     | IntrospectedBaseClass
     | IntrospectedFunction
@@ -16,14 +16,8 @@ export type GirNSMember =
     | IntrospectedEnum
     | IntrospectedAlias;
 
-export const isIntrospectable = (e: { $?: GirInfoAttrs }) =>
-    !e || !e.$ || !e.$.introspectable || e.$.introspectable === "1";
-export const isDeprecated = (e: { $: GirInfoAttrs }) => e && e.$ && e.$.deprecated === "1";
-export const deprecatedVersion = (e: { $: GirInfoAttrs }) => e?.$?.["deprecated-version"];
-export const introducedVersion = (e: { $: GirInfoAttrs }) => e?.$?.version;
-
 export function promisifyNamespaceFunctions(namespace: GirModule) {
-    return namespace.members.forEach(node => {
+    return namespace.members.forEach((node) => {
         if (!(node instanceof IntrospectedFunction)) return;
 
         if (node.parameters.length < 1) return;
@@ -38,16 +32,16 @@ export function promisifyNamespaceFunctions(namespace: GirModule) {
 
         const internal = last_param_unwrapped.type;
 
-        if (internal instanceof TypeIdentifier && internal.is("Gio", "AsyncReadyCallback")) {
+        if (internal instanceof TypeIdentifier && internal.is('Gio', 'AsyncReadyCallback')) {
             const async_res = [
                 ...Array.from(namespace.members.values()).filter(
                     (m): m is IntrospectedFunction => m instanceof IntrospectedFunction
                 )
-            ].find(m => m.name === `${node.name.replace(/_async$/, "")}_finish` || m.name === `${node.name}_finish`);
+            ].find((m) => m.name === `${node.name.replace(/_async$/, '')}_finish` || m.name === `${node.name}_finish`);
 
             if (async_res) {
-                const async_parameters = node.parameters.slice(0, -1).map(p => p.copy());
-                const sync_parameters = node.parameters.map(p => p.copy({ isOptional: false }));
+                const async_parameters = node.parameters.slice(0, -1).map((p) => p.copy());
+                const sync_parameters = node.parameters.map((p) => p.copy({ isOptional: false }));
                 const output_parameters = async_res.output_parameters;
 
                 let async_return = new PromiseType(async_res.return());
@@ -55,10 +49,10 @@ export function promisifyNamespaceFunctions(namespace: GirModule) {
                 if (output_parameters.length > 0) {
                     const raw_return = async_res.return();
                     if (raw_return.equals(VoidType) || raw_return.equals(BooleanType)) {
-                        const [output_type, ...output_types] = output_parameters.map(op => op.type);
+                        const [output_type, ...output_types] = output_parameters.map((op) => op.type);
                         async_return = new PromiseType(new TupleType(output_type, ...output_types));
                     } else {
-                        const [...output_types] = output_parameters.map(op => op.type);
+                        const [...output_types] = output_parameters.map((op) => op.type);
                         async_return = new PromiseType(new TupleType(raw_return, ...output_types));
                     }
                 }
@@ -66,14 +60,14 @@ export function promisifyNamespaceFunctions(namespace: GirModule) {
                 namespace.members.set(node.name, [
                     node.copy({
                         parameters: async_parameters,
-                        return_type: async_return
+                        return_type: async_return,
                     }),
                     node.copy({
-                        parameters: sync_parameters
+                        parameters: sync_parameters,
                     }),
                     node.copy({
-                        return_type: new BinaryType(async_return, node.return())
-                    })
+                        return_type: new BinaryType(async_return, node.return()),
+                    }),
                 ]);
             }
         }
