--- packages/lib/src/gir/nodes.ts
@@ -1,47 +1,0 @@
-export { IntrospectedAlias } from "./alias.ts";
-export {
-    IntrospectedClass,
-    IntrospectedInterface,
-    IntrospectedRecord,
-    GirComplexRecord,
-    IntrospectedBaseClass
-} from "./class.ts";
-export { IntrospectedConstant } from "./const.ts";
-export { IntrospectedEnum, GirEnumMember as IntrospectedEnumMember, IntrospectedError } from "./enum.ts";
-export {
-    IntrospectedFunction,
-    IntrospectedClassCallback,
-    IntrospectedClassFunction,
-    IntrospectedCallback,
-    IntrospectedConstructor,
-    IntrospectedStaticClassFunction,
-    IntrospectedVirtualClassFunction,
-    IntrospectedFunctionParameter,
-    IntrospectedDirectAllocationConstructor as IntrospectedDirectAllocationConstructor
-} from "./function.ts";
-export type { IntrospectedNamespace, GirNSMember } from "./namespace.ts";
-export type { NSRegistry, NSLoader } from "./registry.ts";
-export { IntrospectedField, IntrospectedProperty } from "./property.ts";
-export { IntrospectedSignal, IntrospectedSignalType } from "./signal.ts";
-export {
-    GenericType,
-    TypeExpression,
-    TypeIdentifier,
-    OrType,
-    NativeType,
-    TypeConflict,
-    AnyType,
-    ThisType,
-    TupleType,
-    NeverType,
-    BinaryType,
-    NumberType,
-    ObjectType,
-    StringType,
-    ClosureType,
-    PromiseType,
-    UnknownType,
-    ConflictType
-} from "../gir.ts";
-
-export * as nodeUtils from "./util.ts";
