+++ packages/lib/src/registry.ts
@@ -1,4 +1,4 @@
-import { SanitizedIdentifiers } from './gir/util.ts'
+import { SanitizedIdentifiers } from './utils/index.ts'
 export { FormatGenerator } from './generators/generator.ts'
 
 export * as dts from './generators/dts-modules.ts'
@@ -7,7 +7,7 @@ export * as json from './generators/json.ts'
 
 export * from './generators/index.ts'
 
-export { isInvalid, isPrimitiveType, isSubtypeOf, sanitizeIdentifierName } from './gir/util.ts'
+export { isInvalid, isPrimitiveType, isSubtypeOf, sanitizeIdentifierName } from './utils/index.ts'
 
 export { NSRegistry as GirNSRegistry } from './gir/registry.ts'
 export { Formatter } from './formatters/formatter.ts'
