+++ packages/lib/src/utils/index.ts
@@ -1,8 +1,11 @@
 export * from './documentation.ts'
 export * from './files.ts'
+export * from './gir-parsing.ts'
 export * from './girs.ts'
 export * from './naming.ts'
 export * from './numbers.ts'
 export * from './objects.ts'
 export * from './path.ts'
 export * from './strings.ts'
+export * from './type-resolution.ts'
+export * from './types.ts'
